Fix format-string type mismatch in _CUB_LOG_KERNEL_LAUNCH
Grid/block dims are passed to %d from call sites with mixed signed/unsigned
types. Cast all four to unsigned int and format with %u to match, per
CodeRabbit review on #11584.

+++ cub/cub/detail/logging.cuh
@@ -108,11 +108,11 @@ CUB_NAMESPACE_END
 #ifdef CUB_DEBUG_LOG
 #  define _CUB_LOG_KERNEL_LAUNCH(                                                             \
     kernel_name, grid_dim_x, grid_dim_y, grid_dim_z, block_dim, smem_bytes, stream, fmt, ...) \
-    _CubLog("Invoking " kernel_name "<<<{%d, %d, %d}, %d, %zu, %lld>>>()" fmt "\n",           \
-            grid_dim_x,                                                                       \
-            grid_dim_y,                                                                       \
-            grid_dim_z,                                                                       \
-            block_dim,                                                                        \
+    _CubLog("Invoking " kernel_name "<<<{%u, %u, %u}, %u, %zu, %lld>>>()" fmt "\n",           \
+            static_cast<unsigned int>(grid_dim_x),                                            \
+            static_cast<unsigned int>(grid_dim_y),                                            \
+            static_cast<unsigned int>(grid_dim_z),                                            \
+            static_cast<unsigned int>(block_dim),                                             \
             static_cast<size_t>(smem_bytes),                                                  \
             reinterpret_cast<long long>(stream),                                              \
             ##__VA_ARGS__)
@@ -120,11 +120,11 @@ CUB_NAMESPACE_END
 #  define _CUB_LOG_KERNEL_LAUNCH(                                                             \
     kernel_name, grid_dim_x, grid_dim_y, grid_dim_z, block_dim, smem_bytes, stream, fmt, ...) \
     CUB_NS_QUALIFIER::detail::log(                                                            \
-      "Invoking " kernel_name "<<<{%d, %d, %d}, %d, %zu, %lld>>>()" fmt "\n",                 \
-      grid_dim_x,                                                                             \
-      grid_dim_y,                                                                             \
-      grid_dim_z,                                                                             \
-      block_dim,                                                                              \
+      "Invoking " kernel_name "<<<{%u, %u, %u}, %u, %zu, %lld>>>()" fmt "\n",                 \
+      static_cast<unsigned int>(grid_dim_x),                                                  \
+      static_cast<unsigned int>(grid_dim_y),                                                  \
+      static_cast<unsigned int>(grid_dim_z),                                                  \
+      static_cast<unsigned int>(block_dim),                                                   \
       static_cast<size_t>(smem_bytes),                                                        \
       reinterpret_cast<long long>(stream),                                                    \
       ##__VA_ARGS__)
